week1-appendix/pyproject.toml
[tool.poetry]
name = "api-user"
version = "0.1.0"
description = ""
authors = ["raccoony <raccoonyy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
databases = "^0.6.0"
requests = "^2.27.1"
httpx = "^0.23.0"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.37"}

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
parameterized = "^0.8.1"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
